requirements.txt
Django==5.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==28.4.1
Pillow==10.4.0
python-dotenv==1.0.1
django-cors-headers==4.4.0